# web/doc/swagger.yaml
openapi: 3.0.0
info:
  title: 'Accencture API'
  contact:
    email: vlad.veresklya@yandex.ru
  version: 1.0.0
paths:
...
  /api/warehouse:
    get:
      tags:
        - Склады
      summary: 'Найти все склады'
      parameters:
        -
          name: per-page
          in: query
          description: 'Кол-во записей на странице'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Номер страницы'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Success
  '/api/warehouse/{id}':
    get:
      tags:
        - Склады
      summary: 'Найти склады по id'
      parameters:
        -
          name: id
          in: path
          description: 'ID склада'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
    put:
      tags:
        - Склады
      summary: 'Обновить склад'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                reception_points:
                  description: 'Данные о пунктах приема'
                  type: json
                volume:
                  description: 'Общий объем'
                  type: double
                fullness:
                  description: 'Процент наполненности'
                  type: int
              type: object
      responses:
        '200':
          description: ''
    delete:
      tags:
        - Склады
      summary: 'Удалить склад по id'
      operationId: c2bd7cf0def0c5ed60900743005ee88b
      parameters:
        -
          name: id
          in: path
          description: 'ID склада'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
  /api/warehouse/:
    post:
      tags:
        - Склады
      summary: 'Добавить склад'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                reception_points:
                  description: 'Данные о пунктах приема'
                  type: json
                volume:
                  description: 'Общий объем'
                  type: double
                fullness:
                  description: 'Процент наполненности'
                  type: int
              type: object
      responses:
        '200':
          description: ''
  /api/workers:
    get:
      tags:
        - Работники
      summary: 'Найти всех работников'
      parameters:
        -
          name: per-page
          in: query
          description: 'Кол-во записей на странице'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Номер страницы'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Success
  '/api/workers/{id}':
    get:
      tags:
        - Работники
      summary: 'Найти работника по id'
      parameters:
        -
          name: id
          in: path
          description: 'ID работника'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
  /api/worker/:
    post:
      tags:
        - Работники
      summary: 'Добавить работника'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                name:
                  description: ФИО
                  type: string
                type:
                  description: Тип
                  type: integer
                status:
                  description: Статус
                  type: string
                reception_point:
                  description: 'Пункт приема'
                  type: json
              type: object
      responses:
        '200':
          description: ''
  '/api/worker/{id}':
    put:
      tags:
        - Работники
      summary: 'Обновить работника'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                name:
                  description: ФИО
                  type: string
                type:
                  description: Тип
                  type: integer
                status:
                  description: Статус
                  type: string
                reception_point:
                  description: 'Пункт приема'
                  type: json
              type: object
      responses:
        '200':
          description: ''
    delete:
      tags:
        - Работники
      summary: 'Удалить работника по id'
      operationId: 2875a2f8807f95adb137a01310e95cec
      parameters:
        -
          name: id
          in: path
          description: 'ID работника'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success